создал базовый url, перенёс экшены
