+++ src/components/app/app.jsx
@@ -5,7 +5,7 @@ import AppHeader from './AppHeader/AppHeader.jsx'
 import BurgerIngredients from "./BurgerIngredients/BurgerIngredients";
 import BurgerConstructor from "../BurgerConstructor/BurgerConstructor";
 import { useDispatch, useSelector } from "react-redux";
-import { getData } from "../services/reducers/apiReducer";
+import { getData } from "../services/actions/apiAction";
 import { DndProvider } from 'react-dnd';
 import { HTML5Backend } from 'react-dnd-html5-backend';
