+++ src/components/services/actions/apiAction.jsx
@@ -1,3 +1,5 @@
+import { baseUrl } from "../../../utils/constants";
+
 export const LOAD_START_INGREDIENTS_DATA ='LOAD_START_INGREDIENTS_DATA';
 export const LOAD_START_INGREDIENTS_DATA_REQUEST = 'LOAD_START_INGREDIENTS_DATA_REQUEST';
 export const LOAD_START_INGREDIENTS_DATA_SUCCES = 'LOAD_START_INGREDIENTS_DATA_SUCCES';
@@ -14,4 +16,24 @@ export const removeCount = (ingredient) => {
     return ({type: REMOVE_NUMBER_FROM_COUNT, payload: ingredient})
 }
 
+export const getData = () => {
+    return function(dispatch) {
+        dispatch({
+            type: LOAD_START_INGREDIENTS_DATA_REQUEST
+        })
+        fetch(`${baseUrl}/ingredients`)
+        .then(res => res.json())
+        .then(res => {
+                dispatch({
+                    type: LOAD_START_INGREDIENTS_DATA_SUCCES,
+                    payload: res.data
+                })
 
+        })
+        .catch(err => {
+            dispatch({
+                type: LOAD_START_INGREDIENTS_DATA_FAILED
+            })
+        })
+    }
+}
