+++ src/components/services/actions/orderedIngredientsAction.jsx
@@ -1,3 +1,4 @@
+import { baseUrl } from "../../../utils/constants";
 export const COLLECT_ORDER = 'COLLECT_ORDER';
 export const OPEN_MODAL_WINDOW = 'OPEN_MODAL_WINDOW';
 export const CLOSE_MODAL_WINDOW = 'CLOSE_MODAL_WINDOW';
@@ -5,4 +6,29 @@ export const SEND_ORDER = 'SEND_ORDER';
 
 export const collectOrderGenerator = (ingredients) => {
     return{type: COLLECT_ORDER, payload: ingredients}
+}
+
+export const sendOrder = (ingredients) => {
+    return function(dispatch) {
+
+        fetch(`${baseUrl}/orders`, {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json'
+            },
+            body: JSON.stringify({
+                'ingredients': ingredients.map(ingredients => ingredients._id)
+            })
+        })
+        .then(res => res.json())
+        .then((data) => {
+            dispatch({
+                type: SEND_ORDER,
+                payload: data
+            })
+        })
+        .catch((err) => {
+            console.log( err.status, err.text)
+        })
+    }
 }
