+++ src/components/services/reducers/orderedIngredientsReducer.jsx
@@ -37,28 +37,3 @@ export const orderedIngredientsReducer = (state = initialState, action) => {
         default: return state
     }
 }
-
-export const sendOrder = (ingredients) => {
-    return function(dispatch) {
-
-        fetch('https://norma.nomoreparties.space/api/orders', {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify({
-                'ingredients': ingredients.map(ingredients => ingredients._id)
-            })
-        })
-        .then(res => res.json())
-        .then((data) => {
-            dispatch({
-                type: SEND_ORDER,
-                payload: data
-            })
-        })
-        .catch((err) => {
-            console.log( err.status, err.text)
-        })
-    }
-}
