+++ src/utils/constants.js
@@ -1,0 +1,1 @@
+export const baseUrl = 'https://norma.nomoreparties.space/api'
